pyproject.toml
[tool.poetry]
name = "cms-admin"
version = "0.1.0"
description = ""
authors = ["Zahragrs <dryagaroosi@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4"
psycopg2-binary = ">=2.9.10,<3.0.0"
django = ">=5.2.5,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
django-environ = ">=0.12.0,<0.13.0"
jsonschema = ">=4.25.1,<5.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
django-guardian = ">=3.1.3,<4.0.0"
django-reversion = "^5.1.0"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-beat = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
